refactor: fix impl/unit-test
