+++ include/tools/types/common.hpp
@@ -31,15 +31,15 @@
 #include <tools/types/common/enable_for.hpp>
 #include <tools/types/common/type_of_enum.hpp>
 #include <tools/types/common/add_const_data.hpp>
-#include <tools/types/common/is_dereferencable.hpp>
+//#include <tools/types/common/is_dereferencable.hpp>
 #include <tools/types/common/is_volatile_data.hpp>
 #include <tools/types/common/is_const_data.hpp>
-#include <tools/types/common/is_lambda.hpp>
+//#include <tools/types/common/is_lambda.hpp>
 #include <tools/types/common/array.hpp>
 
-#include <tools/types/common/is_incrementable.hpp>
-#include <tools/types/common/is_decrementable.hpp>
-#include <tools/types/common/is_iterator.hpp>
+//#include <tools/types/common/is_incrementable.hpp>
+//#include <tools/types/common/is_decrementable.hpp>
+//#include <tools/types/common/is_iterator.hpp>
 
 //==============================================================================
 //==============================================================================
