+++ include/tools/types/common/is_const_data.hpp
@@ -1,8 +1,10 @@
 // [2021y-04m-08d][03:02:40] Idrisov Denis R. 100
 // [2021y-04m-10d][07:21:07] Idrisov Denis R. 101
+
+// [2021y-04m-21d][13:09:29] Idrisov Denis R. 102 PRE
 #pragma once
 #ifndef dTOOLS_IS_CONST_DATA_USED_ 
-#define dTOOLS_IS_CONST_DATA_USED_ 101
+#define dTOOLS_IS_CONST_DATA_USED_ 102 PRE
 
 #include <tools/features.hpp>
